+++ lib/domain/models/products_model.dart
