+++ lib/domain/models/user_model.dart
