+++ lib/presentation/blocs/authentication_bloc_screen.dart
@@ -1,6 +1,6 @@
 import 'dart:async';
 
-import 'package:diletta_flutter_test/repository/user_repository.dart';
+import 'package:diletta_flutter_test/data/repository/user_repository.dart';
 import 'package:equatable/equatable.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
