+++ lib/presentation/screen/sign_up_screen.dart
@@ -1,66 +1,80 @@
-import 'package:diletta_flutter_test/blocs/sign_in_bloc.dart';
-import 'package:diletta_flutter_test/component/input_text_field.dart';
-import 'package:diletta_flutter_test/core/input_field_use_case.dart';
-import 'package:diletta_flutter_test/repository/firebase_repository.dart';
+import 'package:diletta_flutter_test/presentation/blocs/sign_up_bloc.dart';
+import 'package:diletta_flutter_test/presentation/component/input_text_field.dart';
+import 'package:diletta_flutter_test/domain/core/input_field_use_case.dart';
+import 'package:diletta_flutter_test/domain/models/user_model.dart';
+import 'package:diletta_flutter_test/data/repository/firebase_repository.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:get_it/get_it.dart';
 
-class SignInScreen extends StatefulWidget {
-  const SignInScreen({super.key});
+class SignUpScreen extends StatefulWidget {
+  const SignUpScreen({super.key});
 
   @override
-  State<SignInScreen> createState() => _SignInScreenState();
+  State<SignUpScreen> createState() => _SignUpScreenState();
 }
 
-class _SignInScreenState extends State<SignInScreen> {
+class _SignUpScreenState extends State<SignUpScreen> {
   final _formKey = GlobalKey<FormState>();
+  final nameController = TextEditingController();
   final emailController = TextEditingController();
   final passwordController = TextEditingController();
   bool obscurePassword = true;
   IconData iconPassword = CupertinoIcons.eye_fill;
-  bool sigInRequired = false;
+  bool signUpRequired = false;
   String? _errorMsg;
 
   @override
   Widget build(BuildContext context) {
     return Form(
-      key: _formKey,
-      child: BlocProvider<SignInBloc>(
-        create: (final context) =>
-            SignInBloc(userRepository: GetIt.I<FirebaseRepository>()),
-        child: BlocListener<SignInBloc, SignInState>(
-          listener: (context, state) {
-            if (state is SigInSuccess) {
-              setState(() {
-                sigInRequired = false;
-              });
-            }
-            if (state is SigInProgress) {
-              setState(() {
-                sigInRequired = true;
-              });
-            }
-            if (state is SigInFailure) {
-              setState(() {
-                sigInRequired = false;
-                _errorMsg = 'Email ou senha inválidos!';
-              });
-            }
-          },
-          child: BlocBuilder<SignInBloc, SignInState>(
-            builder: (final context, final state) {
+        key: _formKey,
+        child: BlocProvider<SignUpBloc>(
+            create: (final context) =>
+                SignUpBloc(userRepository: GetIt.I<FirebaseRepository>()),
+            child: BlocListener<SignUpBloc, SignUpState>(
+                listener: (context, state) {
+              if (state is SignUpSuccessState) {
+                setState(() {
+                  signUpRequired = false;
+                });
+              }
+              if (state is SignUpProcessState) {
+                setState(() {
+                  signUpRequired = true;
+                });
+              }
+              if (state is SignUpFailureState) {
+                setState(() {
+                  signUpRequired = false;
+                  _errorMsg = 'Email já Registrado!';
+                });
+              }
+            }, child: BlocBuilder<SignUpBloc, SignUpState>(
+                    builder: (final context, final state) {
               return Column(
                   mainAxisAlignment: MainAxisAlignment.center,
                   crossAxisAlignment: CrossAxisAlignment.stretch,
                   mainAxisSize: MainAxisSize.min,
                   children: [
                     InputTextField(
+                      controller: nameController,
+                      hintText: 'nome',
+                      inputType: TextInputType.name,
+                      validator: (value) {
+                        if (value!.isEmpty) {
+                          return 'Preencha o campo nome!';
+                        }
+                        return null;
+                      },
+                    ),
+                    const SizedBox(
+                      height: 15,
+                    ),
+                    InputTextField(
                       controller: emailController,
                       hintText: 'email',
                       inputType: TextInputType.emailAddress,
-                      errorMsg: _errorMsg,
                       validator: (value) {
                         if (value!.isEmpty) {
                           return 'Preencha o campo de email!';
@@ -82,6 +96,8 @@ class _SignInScreenState extends State<SignInScreen> {
                         validator: (value) {
                           if (value!.isEmpty) {
                             return 'Preencha o campo de senha';
+                          } else if (passwordRexExp.hasMatch(value)) {
+                            return 'Digite uma senha valida!';
                           }
                           return null;
                         },
@@ -104,7 +120,7 @@ class _SignInScreenState extends State<SignInScreen> {
                     const SizedBox(
                       height: 40,
                     ),
-                    !sigInRequired
+                    !signUpRequired
                         ? SizedBox(
                             height: 50,
                             child: ElevatedButton(
@@ -116,28 +132,31 @@ class _SignInScreenState extends State<SignInScreen> {
                                         .contains(MaterialState.pressed)) {
                                       return Colors.orange;
                                     }
-                                    return Colors.orange;
+                                    return Colors
+                                        .orange; // Use the component's default.
                                   },
                                 ),
                               ),
                               onPressed: () {
                                 if (_formKey.currentState!.validate()) {
-                                  context.read<SignInBloc>().add(
-                                      SignInRequiredEvent(emailController.text,
-                                          passwordController.text));
+                                  UserModel user = UserModel.emptyUser();
+                                  user = user.copyWith(
+                                    name: nameController.text,
+                                    email: emailController.text,
+                                  );
+                                  context.read<SignUpBloc>().add(
+                                      SignUpRequiredEvent(
+                                          user: user,
+                                          password: passwordController.text));
                                 }
                               },
-                              child: const Text('Entrar'),
+                              child: const Text('Cadastrar'),
                             ),
                           )
                         : const Center(
-                          child:
-                              CircularProgressIndicator(color: Colors.orange)),
+                            child: CircularProgressIndicator(
+                                color: Colors.orange)),
                   ]);
-            },
-          ),
-        ),
-      ),
-    );
+            }))));
   }
 }
